modified migrations file to change values to reciept filed to BigIntegerField

+++ bot/migrations/0012_alter_reciept_order_id.py
@@ -13,6 +13,6 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name='reciept',
             name='order_id',
-            field=models.IntegerField(max_length=255),
+            field=models.BigIntegerField(max_length=255),
         ),
     ]
